router: split off sandboxed path from router_darwin
